crates/ngit:intermediate:migrate:update

+++ crates/ngit/test_utils/src/git.rs
@@ -1,19 +1,47 @@
 //create
 
 // implement drop?
-use std::{env::current_dir, fs, path::PathBuf};
+use std::{
+    env::current_dir,
+    fs,
+    path::{Path, PathBuf},
+};
 
 use anyhow::{Context, Result};
-use git2::{Oid, RepositoryInitOptions, Signature, Time};
+use git2::{Branch, Oid, RepositoryInitOptions, Signature, Time};
+use nostr::nips::{nip01::Coordinate, nip19::Nip19Coordinate};
+use nostr_sdk::{Kind, RelayUrl, ToBech32};
+
+use crate::generate_repo_ref_event;
 
 pub struct GitTestRepo {
     pub dir: PathBuf,
     pub git_repo: git2::Repository,
+    pub delete_dir_on_drop: bool,
 }
 
 impl Default for GitTestRepo {
     fn default() -> Self {
-        Self::new("main").unwrap()
+        let repo_event = generate_repo_ref_event();
+        let coordinate = Nip19Coordinate {
+            coordinate: Coordinate {
+                kind: Kind::GitRepoAnnouncement,
+                public_key: repo_event.pubkey,
+                identifier: repo_event.tags.identifier().unwrap().to_string(),
+            },
+            relays: vec![
+                RelayUrl::parse("ws://localhost:8055").unwrap(),
+                RelayUrl::parse("ws://localhost:8056").unwrap(),
+            ],
+        };
+
+        let repo = Self::new("main").unwrap();
+        let _ = repo
+            .git_repo
+            .config()
+            .unwrap()
+            .set_str("nostr.repo", &coordinate.to_bech32().unwrap());
+        repo
     }
 }
 impl GitTestRepo {
@@ -25,9 +53,98 @@ impl GitTestRepo {
                 .initial_head(main_branch_name)
                 .mkpath(true),
         )?;
+        // Make sure we have standard diffs for the tests so that user-level config does
+        // not make them fail.
+        git_repo.config()?.set_bool("diff.mnemonicPrefix", false)?;
         Ok(Self {
             dir: path,
             git_repo,
+            delete_dir_on_drop: true,
+        })
+    }
+    pub fn without_repo_in_git_config() -> Self {
+        Self::new("main").unwrap()
+    }
+
+    pub fn open(path: &PathBuf) -> Result<Self> {
+        let git_repo = git2::Repository::open(path)?;
+        Ok(Self {
+            dir: path.clone(),
+            git_repo,
+            delete_dir_on_drop: true,
+        })
+    }
+
+    pub fn duplicate(existing_repo: &GitTestRepo) -> Result<Self> {
+        let path = current_dir()?.join(format!("tmpgit-{}", rand::random::<u64>()));
+        // function source: https://stackoverflow.com/a/65192210
+        fn copy_dir_all(src: impl AsRef<Path>, dst: impl AsRef<Path>) -> std::io::Result<()> {
+            fs::create_dir_all(&dst)?;
+            for entry in fs::read_dir(src)? {
+                let entry = entry?;
+                let ty = entry.file_type()?;
+                if ty.is_dir() {
+                    copy_dir_all(entry.path(), dst.as_ref().join(entry.file_name()))?;
+                } else {
+                    fs::copy(entry.path(), dst.as_ref().join(entry.file_name()))?;
+                }
+            }
+            Ok(())
+        }
+        copy_dir_all(existing_repo.dir.clone(), path.clone())?;
+        let git_repo = git2::Repository::open(path.clone())?;
+
+        // let git_repo = git2::Repository::clone(existing_repo.dir.to_str().unwrap(),
+        // path.clone())?;
+        Ok(Self {
+            dir: path,
+            git_repo,
+            delete_dir_on_drop: true,
+        })
+    }
+
+    pub fn recreate_as_bare(existing_repo: &GitTestRepo) -> Result<Self> {
+        // create bare repo
+        let path = current_dir()?.join(format!("tmpgit-{}", rand::random::<u64>()));
+        let git_repo = git2::Repository::init_opts(
+            &path,
+            RepositoryInitOptions::new()
+                .initial_head("main")
+                .bare(true)
+                .mkpath(true),
+        )?;
+        // clone existing to a temp repo
+        let tmp_repo = Self::duplicate(existing_repo)?;
+        // add bare as a remote and push branches
+        let mut remote = tmp_repo.git_repo.remote("tmp", path.to_str().unwrap())?;
+        let refspecs = tmp_repo
+            .git_repo
+            .branches(Some(git2::BranchType::Local))?
+            .filter_map(|b| b.ok())
+            .map(|(b, _)| {
+                format!(
+                    "refs/heads/{}:refs/heads/{}",
+                    b.name().unwrap().unwrap(),
+                    b.name().unwrap().unwrap(),
+                )
+            })
+            .collect::<Vec<String>>();
+        remote.push(&refspecs, None)?;
+        // TODO: push tags
+        Ok(Self {
+            dir: path,
+            git_repo,
+            delete_dir_on_drop: true,
+        })
+    }
+
+    pub fn clone_repo(existing_repo: &GitTestRepo) -> Result<Self> {
+        let path = current_dir()?.join(format!("tmpgit-{}", rand::random::<u64>()));
+        let git_repo = git2::Repository::clone(existing_repo.dir.to_str().unwrap(), path.clone())?;
+        Ok(Self {
+            dir: path,
+            git_repo,
+            delete_dir_on_drop: true,
         })
     }
 
@@ -51,6 +168,12 @@ impl GitTestRepo {
         self.stage_and_commit("add t1.md")?;
         fs::write(self.dir.join("t2.md"), "some content1")?;
         self.stage_and_commit("add t2.md")
+    }
+
+    pub fn populate_minus_1(&self) -> Result<Oid> {
+        self.initial_commit()?;
+        fs::write(self.dir.join("t1.md"), "some content")?;
+        self.stage_and_commit("add t1.md")
     }
 
     pub fn populate_with_test_branch(&self) -> Result<Oid> {
@@ -93,10 +216,10 @@ impl GitTestRepo {
         Ok(oid)
     }
 
-    pub fn create_branch(&self, branch_name: &str) -> Result<()> {
+    pub fn create_branch(&self, branch_name: &str) -> Result<Branch> {
         self.git_repo
-            .branch(branch_name, &self.git_repo.head()?.peel_to_commit()?, false)?;
-        Ok(())
+            .branch(branch_name, &self.git_repo.head()?.peel_to_commit()?, false)
+            .context("could not create branch")
     }
 
     pub fn checkout(&self, ref_name: &str) -> Result<Oid> {
@@ -144,7 +267,7 @@ impl GitTestRepo {
         let branch = self
             .git_repo
             .find_branch(branch_name, git2::BranchType::Local)
-            .context(format!("cannot find branch {branch_name}"))?;
+            .context(format!("failed to find branch {branch_name}"))?;
         Ok(branch.into_reference().peel_to_commit()?.id())
     }
 
@@ -152,11 +275,20 @@ impl GitTestRepo {
         self.git_repo.remote(name, url)?;
         Ok(())
     }
+
+    pub fn checkout_remote_branch(&self, branch_name: &str) -> Result<Oid> {
+        self.checkout(&format!("remotes/origin/{branch_name}"))?;
+        let mut branch = self.create_branch(branch_name)?;
+        branch.set_upstream(Some(&format!("origin/{branch_name}")))?;
+        self.checkout(branch_name)
+    }
 }
 
 impl Drop for GitTestRepo {
     fn drop(&mut self) {
-        let _ = fs::remove_dir_all(&self.dir);
+        if self.delete_dir_on_drop {
+            let _ = fs::remove_dir_all(&self.dir);
+        }
     }
 }
 pub fn joe_signature() -> Signature<'static> {
